requirements.txt
#ngsildclient
python-ngsild-client